cleanup on warnings and test results after PR merge relating to multinets

+++ src/spindump_main.c
@@ -84,7 +84,7 @@ int main(int argc,char** argv) {
   FILE* debugfile = stderr;
   struct spindump_main_state* state = spindump_main_initialize();
   interruptFlagLocation = &state->interrupt;
-  srand(clock());
+  srand((unsigned int)clock());
   
   if (state == 0) exit(1);
